move Translator to common module, moved static config part closer to config package for now
